ml/RAG/requirements.txt
# Replicate AI
replicate>=0.15.0

# MongoDB and Vector Search
pymongo>=4.6.0
motor>=3.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# Vector Operations
numpy>=1.24.0
sentence-transformers>=2.2.2

# Logging and Utilities
colorama>=0.4.6
rich>=13.0.0

# Optional: For advanced text processing
nltk>=3.8
spacy>=3.7.0

# HTTP requests (for Replicate API)
requests>=2.31.0